Use new methods for fetching current user information to generate the token
This gets rid of the Exceptions like "DavUserService::getUserId must be of type string, null given"
in cases the stored email address is not really the username.

Also copes with the legacy users which login with an username which is not the email-address.
Use the email-address from the appData container instead.